moved FullTask from src/logic into src/model
